ppt_parser: create new alternative based on records
Sofar, the ppt_parser is rather stupid, does not understand the structure
of the streams but just looks for a certain byte sequence anywhere in the
stream (search_* methods).

There was another attempt to understand and parse the stream structure
but that failed (parse_* methods).

Encouraged by xls_parser, that also parses the data as a series of
records, tried the same with ppt files and works nicely sofar. Might
be able to replace ppt_parser soon.