Potential fix for issue #567
Since rtf and rt can be both destination and reg control word, I
added them to the destination array and added an additional check
in _control_word() to check the char after the control word. If it
is "\" or "{" then it should be treated like a reg control word, if
it is not those chars then it should be a destination control word.