oleobj: Clarify input arg
People (like me) might think that when they give "-r" as argument they
could give a directory with samples as input, e.g.

oleobj -r /path/to/samples

when actually they need to specify that with a glob and -r is
independent of that

oleobj /path/to/samples\*

Clarify that in the help-string of "-r".

Also, giving a glob did result in an error saying that this is not an
existing file. Fixed that.