oleobj: Add --nodump option
oleobj also serves as finder for external relationships. If users want just
that behaviour and no dumping of embedded objects, give them the option to
get that. Also clarify this double-purpose in the doc string.

There is a TODO that might be aimed for exactly that, but I am uncertain
whether I interpret it correctly