Implement detection of plain text
This is not so simple since various text encodings can look rather
"binary", but a few simple heuristics will deal with many text types (at
least those encountered here in Europe).

Of course, all xml is text as well, so use checks for "is this text" only
after more specialized tests like "is this xml".